Start working on intergration with linux & windows
